refactor : 점수 계산 로직을 GameResult 클래스로 분리
